Move variable length field utilities to Varlen.h
