+++ GPL/Events/Helpers.h
@@ -127,96 +127,6 @@ const volatile int consumer_pid = 0;
 // From include/uapi/asm-generic/termbits.h
 #define ECHO 0x00008
 
-// We can't use the ringbuf reserve/commit API if we want to output an event
-// with variable length fields as we won't know the event size in advance, so
-// we create events on the event_buffer_map if this is the case and output them
-// with bpf_ringbuf_output.
-//
-// If the event has no variable length parameters (i.e. is always a fixed
-// size). bpf_ringbuf_reserve/bpf_ringbuf_submit should be used instead to
-// avoid the extra memory copy for better performance.
-
-// 256 KiB per cpu core, 128 KiB is useable as we bail if we go over half
-//
-// Currently, the largest possible event (with variable length parameters all
-// of their respective maximum lengths) is well under 128 KiB. This limit
-// should be bumped if this is no longer the case or else the verifier will
-// start to complain.
-#define EVENT_BUFFER_SIZE (1 << 18)
-#define EVENT_BUFFER_SIZE_HALF (EVENT_BUFFER_SIZE >> 1)
-#define EVENT_BUFFER_SIZE_HALF_MASK (EVENT_BUFFER_SIZE_HALF - 1)
-
-// Convenience macro to determine the current size of an event with its
-// variable length fields
-//
-// We logical and with (EVENT_BUFFER_SIZE - 1). This puts both an upper and
-// lower bound on the value so that we have 0 <= value < EVENT_BUFFER_SIZE, and
-// the verifier is happy.
-#define EVENT_SIZE(x) ((sizeof(*x) + x->vl_fields.size) & (EVENT_BUFFER_SIZE - 1))
-
-// Using a BPF_MAP_TYPE_PERCPU_ARRAY here would be simpler but unfortunately we
-// can't use one. The allocation of map values for a BPF_MAP_TYPE_PERCPU_ARRAY
-// is done under the hood by Linux's percpu allocator, which has a maximum
-// allocation size of 32 KiB (See PCPU_MIN_UNIT_SIZE in include/linux/percpu.h
-// as of Linux 6.0). Trying to create a BPF_MAP_TYPE_PERCPU_ARRAY with a value
-// size larger than 32 KiB will result in -ENOMEM.
-//
-// This is too small for us, so instead we implement a percpu array ourselves
-// using a BPF_MAP_TYPE_ARRAY. We resize it in userspace to $(nproc) elements
-// and access it through the get_event_buffer helper, which returns the value
-// corresponding to the current processor.
-struct {
-    __uint(type, BPF_MAP_TYPE_ARRAY);
-    __uint(key_size, sizeof(int));
-    __uint(value_size, EVENT_BUFFER_SIZE);
-    __uint(max_entries, 0); // Will be resized by userspace to $(nproc)
-} event_buffer_map SEC(".maps");
-
-static void *get_event_buffer()
-{
-    int key = bpf_get_smp_processor_id();
-    return bpf_map_lookup_elem(&event_buffer_map, &key);
-}
-
-// Convenience macro to bail out and submit what we've got if we're out of
-// variable length field space
-//
-// If we're already over half the size of the percpu event buffer, this macro
-// ensures we never add any further fields. If we do, the &
-// EVENT_BUFFER_SIZE_HALF operation below (which is necessary to keep the
-// verifier happy) will cause us to overwrite preceeding variable length
-// fields, resulting in corrupt data
-#define ADD_VL_FIELD_OR_GOTO_EMIT(arg_event, arg_field, arg_type)                                  \
-    do {                                                                                           \
-        if (EVENT_SIZE(arg_event) >= EVENT_BUFFER_SIZE_HALF) {                                     \
-            bpf_printk("Bailing at variable-length field of type %d (event type %d), out of "      \
-                       "buffer space",                                                             \
-                       arg_type, arg_event->hdr.type);                                             \
-            goto emit;                                                                             \
-        }                                                                                          \
-                                                                                                   \
-        struct ebpf_varlen_fields_start *__vlf = &arg_event->vl_fields;                            \
-        struct ebpf_varlen_field *__field =                                                        \
-            (struct ebpf_varlen_field *)(&__vlf->data[__vlf->size & EVENT_BUFFER_SIZE_HALF_MASK]); \
-        __vlf->nfields++;                                                                          \
-        __field->type = arg_type;                                                                  \
-        arg_field     = __field;                                                                   \
-    } while (0)
-
-void ebpf_vl_fields__init(struct ebpf_varlen_fields_start *fields)
-{
-    fields->nfields = 0;
-    fields->size    = 0;
-}
-
-void ebpf_vl_field__set_size(struct ebpf_varlen_fields_start *vl_fields,
-                             struct ebpf_varlen_field *field,
-                             size_t size)
-{
-    vl_fields->size += size + sizeof(struct ebpf_varlen_field);
-    field->size = size;
-}
-
 static bool IS_ERR_OR_NULL(const void *ptr)
 {
     return (!ptr) || (unsigned long)ptr >= (unsigned long)-MAX_ERRNO;
